Remove apidata folder
