+++ core/application/page.go
@@ -8,7 +8,6 @@ import (
 
 	"github.com/Z-M-Huang/Tools/core"
 	"github.com/Z-M-Huang/Tools/core/requestbin"
-	"github.com/Z-M-Huang/Tools/data"
 	"github.com/Z-M-Huang/Tools/utils"
 	"github.com/gin-gonic/gin"
 )
@@ -18,7 +17,7 @@ type Page struct{}
 
 //RenderApplicationPage renders /app/:name
 func (Page) RenderApplicationPage(c *gin.Context) {
-	response := c.Keys[utils.ResponseCtxKey].(*data.Response)
+	response := c.Keys[utils.ResponseCtxKey].(*core.Response)
 
 	name := c.Param("name")
