+++ core/dnslookup/api.go
@@ -7,7 +7,6 @@ import (
 	"strings"
 
 	"github.com/Z-M-Huang/Tools/core"
-	"github.com/Z-M-Huang/Tools/data"
 	"github.com/Z-M-Huang/Tools/utils"
 	"github.com/gin-gonic/gin"
 )
@@ -22,7 +21,7 @@ func (API) DNSLookup(c *gin.Context) {
 
 	err := c.ShouldBind(&request)
 	if err != nil {
-		response.SetAlert(&data.AlertData{
+		response.SetAlert(&core.AlertData{
 			IsDanger: true,
 			Message:  "Invalid lookup request.",
 		})
@@ -37,7 +36,7 @@ func (API) DNSLookup(c *gin.Context) {
 
 	uri, err := url.Parse(request.DomainName)
 	if err != nil {
-		response.SetAlert(&data.AlertData{
+		response.SetAlert(&core.AlertData{
 			IsDanger: true,
 			Message:  "Invalid domain name",
 		})
@@ -46,7 +45,7 @@ func (API) DNSLookup(c *gin.Context) {
 	}
 
 	if uri.Hostname() == "" {
-		response.SetAlert(&data.AlertData{
+		response.SetAlert(&core.AlertData{
 			IsDanger: true,
 			Message:  "Please enter a valid domain name",
 		})
@@ -62,7 +61,7 @@ func (API) DNSLookup(c *gin.Context) {
 	ips, err := net.LookupIP(uri.Hostname())
 	if err != nil {
 		utils.Logger.Error(err.Error())
-		response.SetAlert(&data.AlertData{
+		response.SetAlert(&core.AlertData{
 			IsWarning: true,
 			Message:   "Failed to lookup A records",
 		})
@@ -72,7 +71,7 @@ func (API) DNSLookup(c *gin.Context) {
 			ptrs, err := net.LookupAddr(ip.String())
 			if err != nil {
 				utils.Logger.Error(err.Error())
-				response.SetAlert(&data.AlertData{
+				response.SetAlert(&core.AlertData{
 					IsWarning: true,
 					Message:   fmt.Sprintf("Failed to lookup PTR records for %s", ip.String()),
 				})
@@ -87,7 +86,7 @@ func (API) DNSLookup(c *gin.Context) {
 	cnames, err := net.LookupCNAME(uri.Hostname())
 	if err != nil {
 		utils.Logger.Error(err.Error())
-		response.SetAlert(&data.AlertData{
+		response.SetAlert(&core.AlertData{
 			IsWarning: true,
 			Message:   "Failed to lookup CNAME records",
 		})
@@ -100,7 +99,7 @@ func (API) DNSLookup(c *gin.Context) {
 	nses, err := net.LookupNS(uri.Hostname())
 	if err != nil {
 		utils.Logger.Error(err.Error())
-		response.SetAlert(&data.AlertData{
+		response.SetAlert(&core.AlertData{
 			IsWarning: true,
 			Message:   "Failed to lookup NS records",
 		})
@@ -113,7 +112,7 @@ func (API) DNSLookup(c *gin.Context) {
 	mxes, err := net.LookupMX(uri.Hostname())
 	if err != nil {
 		utils.Logger.Error(err.Error())
-		response.SetAlert(&data.AlertData{
+		response.SetAlert(&core.AlertData{
 			IsWarning: true,
 			Message:   "Failed to lookup MX records",
 		})
@@ -126,7 +125,7 @@ func (API) DNSLookup(c *gin.Context) {
 	txts, err := net.LookupTXT(uri.Hostname())
 	if err != nil {
 		utils.Logger.Error(err.Error())
-		response.SetAlert(&data.AlertData{
+		response.SetAlert(&core.AlertData{
 			IsWarning: true,
 			Message:   "Failed to lookup TXT records",
 		})
