+++ core/hilosimulator/api.go
@@ -2,7 +2,6 @@ package hilosimulator
 
 import (
 	"github.com/Z-M-Huang/Tools/core"
-	"github.com/Z-M-Huang/Tools/data"
 	"github.com/Z-M-Huang/Tools/utils"
 	hilosimulator "github.com/Z-M-Huang/hilosimulator"
 	"github.com/gin-gonic/gin"
@@ -19,7 +18,7 @@ func (API) HILOSimulate(c *gin.Context) {
 	err := c.ShouldBind(&request)
 	if err != nil {
 		utils.Logger.Error(err.Error())
-		response.SetAlert(&data.AlertData{
+		response.SetAlert(&core.AlertData{
 			IsDanger: true,
 			Message:  "Invalid simulation request.",
 		})
@@ -28,14 +27,14 @@ func (API) HILOSimulate(c *gin.Context) {
 	}
 
 	if request.RollAmount < 0 {
-		response.SetAlert(&data.AlertData{
+		response.SetAlert(&core.AlertData{
 			IsWarning: true,
 			Message:   "Roll Amount: Cannot be negative number",
 		})
 		core.WriteResponse(c, 400, response)
 		return
 	} else if request.RollAmount > 50000 {
-		response.SetAlert(&data.AlertData{
+		response.SetAlert(&core.AlertData{
 			IsDanger: true,
 			Message:  "Requested Roll Amount is too large. Please do batches and keep the server health. Thank you",
 		})
@@ -70,7 +69,7 @@ func (API) HILOSimulate(c *gin.Context) {
 	result, err := hilosimulator.Simulate(simConfig)
 	if err != nil {
 		utils.Logger.Error(err.Error())
-		response.SetAlert(&data.AlertData{
+		response.SetAlert(&core.AlertData{
 			IsDanger: true,
 			Message:  err.Error(),
 		})
@@ -90,7 +89,7 @@ func (API) HILOVerify(c *gin.Context) {
 	err := c.ShouldBind(&request)
 	if err != nil {
 		utils.Logger.Error(err.Error())
-		response.SetAlert(&data.AlertData{
+		response.SetAlert(&core.AlertData{
 			IsDanger: true,
 			Message:  "Invalid simulation request.",
 		})
@@ -101,7 +100,7 @@ func (API) HILOVerify(c *gin.Context) {
 	valid, err := hilosimulator.Verify(request.ClientSeed, request.ServerSeed, request.Nonce, request.Roll)
 	if err != nil {
 		utils.Logger.Error(err.Error())
-		response.SetAlert(&data.AlertData{
+		response.SetAlert(&core.AlertData{
 			IsDanger: true,
 			Message:  err.Error(),
 		})
