+++ core/kellycriterion/api.go
@@ -4,7 +4,6 @@ import (
 	"math"
 
 	"github.com/Z-M-Huang/Tools/core"
-	"github.com/Z-M-Huang/Tools/data"
 	"github.com/Z-M-Huang/Tools/utils"
 	kellycriterion "github.com/Z-M-Huang/kelly-criterion"
 	"github.com/gin-gonic/gin"
@@ -15,12 +14,12 @@ type API struct{}
 
 //Simulate /api/kelly-criterion/simulate
 func (API) Simulate(c *gin.Context) {
-	response := c.Keys[utils.ResponseCtxKey].(*data.Response)
+	response := c.Keys[utils.ResponseCtxKey].(*core.Response)
 	var simulationResult []*Response
 	request := &Request{}
 	err := c.ShouldBind(&request)
 	if err != nil {
-		response.SetAlert(&data.AlertData{
+		response.SetAlert(&core.AlertData{
 			IsDanger: true,
 			Message:  "Invalid simulation request.",
 		})
