--- data/apidata/application/request_bin.go
@@ -1,1 +1,0 @@
-package application
