--- data/apidata/login_response.go
@@ -1,7 +1,0 @@
-package apidata
-
-//LoginResponse login response
-type LoginResponse struct {
-	IsSuccess bool
-	Redirect  string
-}
