--- data/webdata/account_page_data.go
@@ -1,6 +1,0 @@
-package webdata
-
-//AccountPageData /account
-type AccountPageData struct {
-	HasPassword bool
-}
